[model] - Create a SimpleModel in order to clean BaseModel
Changes:
- BaseModel just preserve the solve function
- SimpleModel implement distance and merge
- model.py renamed to positions.py
- more_popular.py renamed to most_popular.py
- All models except DEA/AR change to inherit from SimpleModel
