+++ src/model/__init__.py
@@ -1,7 +1,7 @@
 from .base import ModelError
-from .model import Positions
+from .positions import Positions
 from .minimun import Minimum
-from .more_popular import MostPopular
+from .most_popular import MostPopular
 from .dea_ar_model import DEA_AR
 
 def use_model(data, model_name=None):
