+++ src/model/minimun.py
@@ -1,4 +1,4 @@
-from .model import Positions
+from .positions import Positions
 
 class Minimum(Positions):
     '''Busca el calendario que minimice la máxima cantidad de cambios de orden con respecto a las propuestas'''
